+++ Libplanet/Action/ActionContext.cs
@@ -6,11 +6,13 @@ namespace Libplanet.Action
             Address signer,
             long blockIndex,
             IAccountStateDelta previousStates,
-            int randomSeed
+            int randomSeed,
+            bool rehearsal = false
         )
         {
             Signer = signer;
             BlockIndex = blockIndex;
+            Rehearsal = rehearsal;
             PreviousStates = previousStates;
             Random = new Random(randomSeed);
         }
@@ -18,6 +20,8 @@ namespace Libplanet.Action
         public Address Signer { get; }
 
         public long BlockIndex { get; }
+
+        public bool Rehearsal { get; }
 
         public IAccountStateDelta PreviousStates { get; }