+++ Libplanet/Action/IActionContext.cs
@@ -19,6 +19,14 @@ namespace Libplanet.Action
         long BlockIndex { get; }
 
         /// <summary>
+        /// Whether an <see cref="IAction"/> is being executed during
+        /// &#x201c;rehearsal mode&#x201d;, that there is nothing
+        /// in <see cref="PreviousStates"/>.
+        /// </summary>
+        /// <seealso cref="Libplanet.Tx.Transaction{T}.Create"/>
+        bool Rehearsal { get; }
+
+        /// <summary>
         /// A null delta of states, which means it represents the states
         /// before <see cref="IAction"/> executes.
         /// <para>Although a <see cref="IAccountStateDelta"/> instance is